feat(vision): 피듀셜 서브픽셀 보정 및 좌표 소수 전송, Docker 주석 반영
- Edge: YOLO 후 ROI 타원 피팅 서브픽셀, 정렬·정합에 center_x/y_subpx 사용
- Edge: InspectionPacket fiducial 좌표 float, main에서 소수 반올림 전송
- Spring/.NET: fiducial X/Y를 Double로 수신·저장·응답
- Dockerfile: OpenCV 용도 및 DB DOUBLE 주석

+++ backend-dotnet/src/AiFactory.Api/Contracts/InspectionIngestRequest.cs
@@ -12,16 +12,16 @@ public sealed class InspectionIngestRequest
     public InspectionStatus Result { get; set; }
 
     [JsonPropertyName("fiducial1X")]
-    public int? Fiducial1X { get; set; }
+    public double? Fiducial1X { get; set; }
 
     [JsonPropertyName("fiducial1Y")]
-    public int? Fiducial1Y { get; set; }
+    public double? Fiducial1Y { get; set; }
 
     [JsonPropertyName("fiducial2X")]
-    public int? Fiducial2X { get; set; }
+    public double? Fiducial2X { get; set; }
 
     [JsonPropertyName("fiducial2Y")]
-    public int? Fiducial2Y { get; set; }
+    public double? Fiducial2Y { get; set; }
 
     [JsonPropertyName("fiducial1Confidence")]
     public double? Fiducial1Confidence { get; set; }
